requirements.txt
# Stock AI Investor - 依存ライブラリ
# pip install -r requirements.txt

# データ取得・処理
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.3
requests==2.31.0

# テクニカル分析
ta==0.11.0
pandas-ta==0.3.14b

# 可視化
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# 設定管理
pyyaml==6.0.1
python-dotenv==1.0.0

# 通知
discord-webhook==1.3.0

# スケジューリング
schedule==1.2.1

# ログ
loguru==0.7.2

# テスト
pytest==7.4.4
pytest-cov==4.1.0
